+++ apps/supplement-stack/CHANGELOG.md
@@ -1,5 +1,17 @@
 # supplement-stack
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - zemble-plugin-auth-otp@0.0.10
+  - @zemble/core@0.0.8
+  - @zemble/graphql@0.0.12
+  - zemble-plugin-apple-app-site-association@0.0.6
+  - zemble-plugin-bull@0.0.10
+  - zemble-plugin-todo@0.0.8
+
 ## 0.0.7
 
 ### Patch Changes