+++ packages/apple-app-site-association/CHANGELOG.md
@@ -1,5 +1,12 @@
 # zemble-plugin-apple-app-site-association
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/core@0.0.8
+
 ## 0.0.5
 
 ### Patch Changes