+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # zemble-plugin-auth
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/core@0.0.8
+  - @zemble/graphql@0.0.12
+
 ## 0.0.9
 
 ### Patch Changes