+++ packages/cms/CHANGELOG.md
@@ -1,5 +1,14 @@
 # zemble-plugin-cms
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/core@0.0.8
+  - @zemble/graphql@0.0.12
+  - zemble-plugin-auth@0.0.10
+
 ## 0.0.7
 
 ### Patch Changes