+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @zemble/core
 
+## 0.0.8
+
+### Patch Changes
+
+- Fix dependency resolution
+
 ## 0.0.7
 
 ### Patch Changes