+++ packages/email-sendgrid/CHANGELOG.md
@@ -1,5 +1,14 @@
 # zemble-plugin-email-sendgrid
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/core@0.0.8
+  - @zemble/graphql@0.0.12
+  - zemble-plugin-kv@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes