+++ packages/kv/CHANGELOG.md
@@ -1,5 +1,14 @@
 # zemble-plugin-kv
 
+## 0.0.10
+
+### Patch Changes
+
+- Fix dependency resolution
+- Updated dependencies
+  - @zemble/core@0.0.8
+  - @zemble/graphql@0.0.12
+
 ## 0.0.9
 
 ### Patch Changes