+++ packages/logger-graphql/CHANGELOG.md
@@ -1,5 +1,13 @@
 # zemble-plugin-logger-graphql
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/core@0.0.8
+  - @zemble/graphql@0.0.12
+
 ## 0.0.7
 
 ### Patch Changes